requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cn2an==0.5.19
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
Cython==0.29.33
datasets==2.10.1
decorator==5.1.1
dill==0.3.6
emoji==1.7.0
filelock==3.10.0
frozenlist==1.3.3
fsspec==2023.3.0
gevent==22.10.2
greenlet==2.0.2
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
MarkupSafe==2.1.2
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
numba==0.56.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
playsound==1.2.2
pooch==1.6.0
preshed==3.0.8
proces==0.1.4
pyarrow==11.0.0
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.6
pydub==0.25.1
pyopenjtalk==0.3.0
pypinyin==0.48.0
pysentimiento==0.6.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
romajitable==0.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soxr==0.3.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.27.2
typer==0.7.0
typing_extensions==4.5.0
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.15
wasabi==1.1.1
websocket-client==1.5.1
websockets==10.4
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
